requirements.txt
absl-py==1.4.0
astunparse==1.6.3
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.2.0
contextlib2==21.6.0
flatbuffers==23.5.26
gast==0.4.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.56.2
h5py==3.9.0
idna==3.4

numpy==1.24.3
opencv-python==4.8.0.74
Pillow==10.0.0
protobuf==4.23.4
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.32.0
tf-slim==1.1.0
contextlib2==21.6.0